+++ libs/importer/src/importers/lastpass/lib/chunk.ts
@@ -1,0 +1,4 @@
+export class Chunk {
+  id: string;
+  payload: Uint8Array;
+}
