+++ libs/importer/src/importers/lastpass/lib/otpMethod.ts
@@ -1,0 +1,5 @@
+export enum OtpMethod {
+  GoogleAuth,
+  MicrosoftAuth,
+  Yubikey,
+}
